requirements.txt
Django==5.0.2
django-decouple==2.1
djangorestframework==3.14.0
openpyxl==3.1.2
